# src/factories/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Koa Typescript Restfull API",
        "description": "This is a technical Demo for a restfull API server on node.js with typescript and Koa server.",
        "contact": {
            "name": "Philippe Fouque",
            "url": "http://www.pfouque.fr"
        }
    },
    "externalDocs": {
        "description": "github repository",
        "url": "https://github.com/pfouque06/koa-typescript-restfull-api"
    },
    "host": "localhost",
    "basePath": "",
    "schemes": ["https","http"],
    "tags": [
        {
            "name": "hello",
            "description": "Hello operation"
        },{
            "name": "auth",
            "description": "Operations about authentication"
        },{
            "name": "users",
            "description": "Operations about user"
        }
    ],
    "paths": {
        "": {
            "get":{
                "tags": ["hello"],
                "summary": "Hello",
                "operationId": "Hello",
                "security": [],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/register": {
            "post":{
                "tags": ["auth"],
                "summary": "register a new user with its credentials (email, password)",
                "operationId": "register",
                "security": [],
                "requestBody": {
                    "description": "user Credential: email, password",
                    "required": "true",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/LoginForm"
                            },
                            "examples": {
                                "Sam Plai": {
                                    "summary": "credentials for user Sam Plai",
                                    "value": {
                                        "email": "sam.plai@gmail.com",
                                        "password": "secret"
                                    }
                                },
                                "Sam Bote": {
                                    "summary": "credentials for user Sam Bote",
                                    "value": {
                                        "email": "sam.bote@gmail.com",
                                        "password": "secret"
                                    }
                                }
                            }
                        }
                    }
                },
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/login": {
            "post":{
                "tags": ["auth"],
                "summary": "login with user credentials: (email, password)",
                "operationId": "login",
                "security": [],
                "requestBody": {
                    "description": "user Credential: email, password",
                    "required": "true",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/LoginForm"
                            },
                            "examples": {
                                "admin account": {
                                    "summary": "credentials for admin Sam Va",
                                    "value": {
                                        "email": "sam.va@gmail.com",
                                        "password": "secret"
                                    }
                                },
                                "Sam Plai": {
                                    "summary": "credentials for user Sam Plai",
                                    "value": {
                                        "email": "sam.plai@gmail.com",
                                        "password": "secret"
                                    }
                                },
                                "Sam Bote": {
                                    "summary": "credentials for user Sam Bote",
                                    "value": {
                                        "email": "sam.bote@gmail.com",
                                        "password": "secret"
                                    }
                                }
                            }
                        }
                    }
                },
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/logout": {
            "post":{
                "tags": ["auth"],
                "summary": "logout current User",
                "operationId": "logout",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/test": {
            "get":{
                "tags": ["auth"],
                "summary": "test api access of current User",
                "operationId": "testUserAccess",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/reset": {
            "post":{
                "tags": ["auth"],
                "summary": "reset JWT redis db",
                "operationId": "resetRedis",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/users":{
            "get":{
                "tags": ["users"],
                "summary": "Get all users",
                "description": "Restricted to logged user.",
                "operationId": "getUsers",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/User"
                                    }
                                }
                            }
                        }

                    },
                    "404": {
                        "description": "Invalid username/password supplied"
                    }
                }
            },
            "post": {
                "tags": ["users"],
                "summary": "create a new user",
                "operationId": "postUser",
                "requestBody": {
                    "description": "user Type: email, password are required",
                    "required": "true",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/User"
                            },
                            "examples": {
                                "user account example": {
                                    "summary": "credentials for user Sam Plai",
                                    "value": {
                                        "profile": "user",
                                        "firstName": "Sam",
                                        "lastName": "Plai",
                                        "mobile": "0123456789",
                                        "email": "sam.plai@gmail.com",
                                        "password": "secret"
                                    }
                                }
                            }
                        }
                    }
                },
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/users/{id}":{
            "get":{
                "tags": ["users"],
                "summary": "Get specific user by its id",
                "description": "Restricted to logged user.",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": "true",
                        "description": "id of user",
                        "schema": {
                            "type": "number",
                            "minimum": "1"
                        }
                    }
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid username/password supplied"
                    }
                }
            },
            "put": {
                "tags": ["users"],
                "summary": "update a user",
                "operationId": "putUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": "true",
                        "description": "id of user",
                        "schema": {
                            "type": "number",
                            "minimum": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "user Type: email, password are required",
                    "required": "true",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/User"
                            },
                            "examples": {
                                "Sam Plai": {
                                    "summary": "credentials for user Sam Plai",
                                    "value": {
                                        "profile": "user",
                                        "firstName": "Sam",
                                        "lastName": "Plai",
                                        "mobile": "0123456789",
                                        "email": "sam.plai@gmail.com",
                                        "password": "secret"
                                    }
                                }
                            }
                        }
                    }
                },
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid username/password supplied"
                    }
                }
            },
            "delete": {
                "tags": ["users"],
                "summary": "delete a user",
                "operationId": "delUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": "true",
                        "description": "id of user",
                        "schema": {
                            "type": "number",
                            "minimum": "1"
                        }
                    }
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/users/reset":{
            "post": {
                "tags": ["users"],
                "summary": "reset user and associaed tables",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        }
    },
    "definitions" : {
        "User": {
            "type": "object",
            "required": [ "email", "password" ],
            "properties": {
                "id": {
                    "type": "number"
                },
                "firstName":{
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "birthDate": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "profile": {
                    "type": "string",
                    "enum": ["user", "admin"]
                }
            }
        },
        "LoginForm": {
            "type": "object",
            "required": [ "email", "password" ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "description": "JWT Authorization",
                "type": "http",
                "scheme": "bearer",
                "in": "in",
                "bearerFormat": "JWT" 
            }
        }
    },
    "security": [
        { "bearerAuth": [] }
    ]
}